Add simple model (which compiles!) to file
I added a simple model with just `Age` as a predictor so that we could compile something. Before fitting the model, I made a graph for guidance. It would be good to explore this simple model further, including a prior predictive check and a posterior predictive check. At very least, there should be a graph of the model. (Maybe I'll add that next.)