Add some explanation for simple model and graph of posterior fit
I added a graph of the posterior fit for the simple model that I proposed. (It's bad.) There may be a package (tidybayes?) that automates making this sort of graph. I planned to make a graph on the probability scale (instead of logit scale), but, as far as I could tell, geom_function only plots one function at a time. The geom_abline layer, on the other hand takes vectors of slopes and intercepts.