Ran the model
Ran the big model and saved the RDS locally becasue GitHub won't let me upload it. I will use it to reference my figures which will mean that only I can view it, so we may have to find some way that I can transport it to Dr. Parson if he wants to take a peek.